Move entities under the intern namespace
