+++ lib/entities/sys/vote.js
@@ -1,5 +1,5 @@
 const Assert = require('assert-plus')
-const { fetchProp } = require('../../lib/utils')
+const { fetchProp } = require('../../utils')
 
 class Vote {
   static TYPE_UP() {
