+++ services/get_vote_stats/index.js
@@ -1,5 +1,5 @@
 const Assert = require('assert-plus')
-const Vote = require('../../entities/sys/vote')
+const Vote = require('../../lib/entities/sys/vote')
 const groupBy = require('lodash.groupby')
 const { fetchProp, countMatching } = require('../../lib/utils')
